backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart
langchain
langchain-openai
langchain-google-genai
scikit-learn
pandas

bleak
aiohttp
